Full rebuild HomeWork after lesson 2. Ready to HW lesson 3.
