+++ VK/FriendsCollectionView.swift
@@ -1,0 +1,29 @@
+
+import UIKit
+
+private let reuseIdentifier = "Cell"
+
+class FriendsCollectionView: UICollectionViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        print("[Logging] load Friends Photo Collection View")
+        self.collectionView!.register(UICollectionViewCell.self, forCellWithReuseIdentifier: reuseIdentifier)
+
+    }
+
+    override func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return 0
+    }
+
+
+    override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 0
+    }
+
+    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: reuseIdentifier, for: indexPath)
+    
+        return cell
+    }
+}
