+++ VK/GroupInfoTableView.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+class GroupInfoTableView: UITableViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        return 0
+    }
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 0
+    }
+}
