+++ VK/LoginForm.swift
@@ -1,4 +1,4 @@
-//Complete homework , lesson 1 & 2
+//Complete homework , lesson 3
 import UIKit
 
 class LoginForm: UIViewController {
@@ -18,6 +18,7 @@ class LoginForm: UIViewController {
         let login = loginForm.text!
         print("[Logging] try to connect with username:\(login)")
     }
+
     override func shouldPerformSegue(withIdentifier identifier: String, sender: Any?) -> Bool {
         let login = loginForm.text!
         let password = passwordForm.text!
@@ -67,7 +68,9 @@ class LoginForm: UIViewController {
     }
     // функция для возврата на данный экран , объявляем там куда хотим вернуться
     // button -> "Exit" on View upSettings, использована в реализации кнопки - Logout
-    @IBAction func goToLofinForm (unwindSegue: UIStoryboardSegue) {}
+    @IBAction func goToLofinForm (unwindSegue: UIStoryboardSegue) {
+        print("[Logging] User \(loginForm.text ?? "Unknown") disconnected")
+    }
     
     //Реализация метода фиксации Segue ( прыгаем по переходу Profile Form )
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -91,147 +94,3 @@ class LoginForm: UIViewController {
         view.endEditing(true)
     }
 }
-
-class ProfileForm: UITabBarController {
-    
-    var username = ""
-    var language = 0
-    
-    //реализация метода получения данных с предыдущего View
-    override func viewDidLoad() {
-        print("[Logging] connect with username - \(username)")
-        if language == 0 {
-            print("[Logging] connect with language option Ru")
-        } else {
-            print("[Logging] connect with language option En")
-        }
-    }
-    @objc func hideKeyboard() {
-        view.endEditing(true)
-    }
-}
-var dataFriends: [String] = ["Friend_One", "Friend_Two", "Friend_Three", "Friend_Four", "Friend_Five"]
-
-class FriendsTableView: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    
-    //реализация количества строк (ячеек) равное количеству элементов массива dataFriens
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dataFriends.count
-    }
-    //реализация присвоения титулу ячеек значений элементов массива data, идентификатор CellFriends задается в Storyboard
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "CellFriends", for: indexPath)
-        cell.textLabel?.text = dataFriends[indexPath.row]
-        return cell
-    }
-    //реализация функции при нажатии на Cell
-    //let main =  UIStoryboard()
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        //плавная анимация исчезновения выделения
-        tableView.deselectRow(at: indexPath, animated: true)
-        //выведем в консоль имя нажатой ячейки
-        print(dataFriends[indexPath.row])
-        //сделаем переключение на Collection View, со всех ячеек переключаемся на один и тот же Collection View.
-        let main = UIStoryboard( name: "Main", bundle: nil)
-        let vc = main.instantiateViewController(identifier: "PhotoFreindsCollection")
-        navigationController?.pushViewController(vc, animated: true)
-        
-    }
-    
-    @IBOutlet var friendsTView: UITableView!
-    
-    var username = ""
-    //реализация метода получения данных с предыдущего View
-    override func viewDidLoad() {
-        print("[Logging] load Friends & Photo View")
-    }
-    @objc func hideKeyboard() {
-        view.endEditing(true)
-    }
-}
-
-var dataGroups: [String] = ["Group_One", "Group_Two", "Group_Three", "Group_Four", "Group_Five"]
-
-class GroupsTableView: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    
-    @IBOutlet var GroupsView: UITableView!
-    //реализация количества строк (ячеек) равное количеству элементов массива dataGroups
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dataGroups.count
-    }
-    //реализация присвоения титулу ячеек значений элементов массива dataGroups, идентификатор CellGroups задается в Storyboard
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "CellGroups", for: indexPath)
-        cell.textLabel?.text = dataGroups[indexPath.row]
-        return cell
-    }
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        //плавная анимация исчезновения выделения
-        tableView.deselectRow(at: indexPath, animated: true)
-        //выведем в консоль имя нажатой ячейки
-        print(dataGroups[indexPath.row])
-        //сделаем переключение на alert - Error! так как у нас пока нет внутренностей для групп
-        let alert = UIAlertController(title: "Error", message: "Access error", preferredStyle: .alert)
-        let action = UIAlertAction(title: "OK", style: .cancel, handler: nil)
-        alert.addAction(action)
-        present(alert, animated: true, completion: nil)
-    }
-    
-    override func viewDidLoad() {
-        print("[Logging] load Groups View")
-    }
-    @objc func hideKeyboard() {
-        view.endEditing(true)
-    }
-}
-
-var dataGlobalGroups: [String] = ["Swift Develop Group 1", "Swift Develop Group 2", "Swift Develop Group 3", "Swift Develop Group 4" , "Swift Develop Group 5", "Swift Develop Group 6", "Swift Develop Group 7", "Swift Develop Group 8", "Swift Develop Group 9"]
-
-class GlobalSearchGroupsTableView: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    
-    @IBOutlet var GlobalSearchGroupView: UITableView!
-    
-    //реализация количества строк (ячеек) равное количеству элементов массива dataGroups
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dataGlobalGroups.count
-    }
-    //реализация присвоения титулу ячеек значений элементов массива dataGroups, идентификатор CellGlobalGroups задается в Storyboard
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "CellGlobalGroups", for: indexPath)
-        cell.textLabel?.text = dataGlobalGroups[indexPath.row]
-        return cell
-    }
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        //плавная анимация исчезновения выделения
-        tableView.deselectRow(at: indexPath, animated: true)
-        //выведем в консоль имя нажатой ячейки
-        print(dataGlobalGroups[indexPath.row])
-        //сделаем переключение на alert - Error! так как у нас пока нет внутренностей для групп
-        let alert = UIAlertController(title: "Error", message: "Access error", preferredStyle: .alert)
-        let action = UIAlertAction(title: "OK", style: .cancel, handler: nil)
-        alert.addAction(action)
-        present(alert, animated: true, completion: nil)
-    }
-
-    override func viewDidLoad() {
-        print("[Logging] load Global Search Groups View")
-    }
-    @objc func hideKeyboard() {
-        view.endEditing(true)
-    }
-}
-
-class FriendsPhotoCollectionView: UICollectionViewController {
-    override func viewDidLoad() {
-        print("[Logging] load Friends Photo Collection View")
-    }
-}
-class GroupInfoTableView: UITableViewController {
-    override func viewDidLoad() {
-        print("[Logging] load Group Info Table View")
-    }
-}
-// делаем кучу всяких разных действий
-// что бы залить это дело на гит хаб ...
-// ну и еще разок ...
-// и еще ...
